DonateButton.jsx refactored to fire on user click
